requirements.txt
feedparser>=6.0.10
openpyxl>=3.1.2
python-pptx>=0.6.23
requests>=2.31.0
